requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# LLM + RAG
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.8
openai==1.30.1

# Vector database
faiss-cpu==1.8.0
tiktoken==0.7.0

# PDF processing
pypdf==4.2.0

# Config & utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
